Updated files - Redesigned configuration interface - Added new configuration design, no left panel
